requirements.txt
openai==0.28
streamlit
python-dotenv
fpdf
replicate
requests